Changing functions return

+++ src/usecases/favorites-usecases.ts
@@ -1,20 +1,28 @@
 import { FavoriteRepositoryPrisma } from "../repositories/favorites-repository";
-import { AuthService } from "../services/auth-service";
 import {
   AddFavorite,
-  Favorite,
   FavoriteRepository,
   GetFavorite,
 } from "../interfaces/favorites-interface";
+import { ProjectRepositoryPrisma } from "../repositories/projects-repository";
+import { ProjectRepository } from "../interfaces/projects-interface";
 
 class FavoriteUseCase {
   private favoriteRepository: FavoriteRepository;
+  private projectRepository: ProjectRepository;
 
   constructor() {
     this.favoriteRepository = new FavoriteRepositoryPrisma();
+    this.projectRepository = new ProjectRepositoryPrisma();
   }
 
   async addFavorite({ userId, projectId }: AddFavorite): Promise<void> {
+    const project = await this.projectRepository.findProjectById(projectId);
+
+    if (!project.project) {
+      throw new Error("Project not found.");
+    }
+
     await this.favoriteRepository.addFavorite({
       userId,
       projectId,
@@ -22,6 +30,12 @@ class FavoriteUseCase {
   }
 
   async deleteFavorite({ userId, projectId }: AddFavorite): Promise<void> {
+    const project = await this.projectRepository.findProjectById(projectId);
+
+    if (!project.project) {
+      throw new Error("Project not found.");
+    }
+
     await this.favoriteRepository.deleteFavorite({ userId, projectId });
   }
